Rename pass. Website everything. We are not a web browser lol. infoEl, infoElement, infoList. websiteURL, websiteHTML, command, button, field. words that describe, everything was called website.
Begin to implement json edit and delete.
Dropped checkboxes, we are going to move into longpress

got rid of a loop pouring infoElements into the UI.
now it delivers a truck of infoElements, for the UI to grab.
that fixes duplicates issue. As infoEls are not going to be repoured.