Fixed big websites having problems with displaying
Previous commit Fixed freeze and crash on large websites, by moving them into backThread.
But large websites still weren't working.

Turns out cat was waiting for space in the buffer.
And buffer reader was waiting for cat to exit, before it starts reading.

Having no wait before buffer reading, makes the buffer block to hold for the process to write, but reads after the process writes, so process doesn't need to wait. And gets to complete.